% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arrow_notation}
\alias{arrow_notation}
\title{Arrow notation}
\format{
A vector of notational symbols that provides an arrow separator ("a -> b")
between prefix and suffix.
}
\usage{
arrow_notation
}
\description{
A description of arrow notation.
}
\examples{
arrow_notation
}
\keyword{datasets}
