% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bracket_arrow_notation}
\alias{bracket_arrow_notation}
\title{Bracket arrow notation}
\format{
A vector of notational symbols that provides bracket arrow ("a [-> b]") notation.
}
\usage{
bracket_arrow_notation
}
\description{
A description of bracket arrow notation.
}
\examples{
bracket_arrow_notation
}
\keyword{datasets}
