% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bracket_notation}
\alias{bracket_notation}
\title{Bracket notation}
\format{
A vector of notational symbols that provides bracket ("a [b]") notation.
}
\usage{
bracket_notation
}
\description{
A description of bracket notation.
}
\examples{
bracket_notation
}
\keyword{datasets}
