% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dash_notation}
\alias{dash_notation}
\title{A description of dash notation.}
\format{
A vector of notational symbols that provides an dash separator ("a - b")
between prefix and suffix.
}
\usage{
dash_notation
}
\description{
A description of dash notation.
}
\examples{
dash_notation
}
\keyword{datasets}
