% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{first_dot_notation}
\alias{first_dot_notation}
\title{First dot notation}
\format{
A vector of notational symbols that provides first dot ("a.b") notation.
}
\usage{
first_dot_notation
}
\description{
A description of first dot notation.
Note that "a.b.c" splits into prefix ("a") and suffix ("b.c").
}
\examples{
first_dot_notation
}
\keyword{datasets}
