% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{from_notation}
\alias{from_notation}
\title{From notation}
\format{
A vector of notational symbols that provides from ("a [from b]") notation.
}
\usage{
from_notation
}
\description{
A description of from notation.
}
\examples{
from_notation
}
\keyword{datasets}
