% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_noun}
\alias{get_noun}
\title{Extract nouns from labels}
\usage{
get_noun(labels, notation)
}
\arguments{
\item{notation}{}
}
\value{
The nouns from row and column labels.
}
\description{
Extract nouns from labels
}
\examples{
get_noun("a [b]", bracket_notation)
}
