% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_piece}
\alias{get_piece}
\title{Get a piece of a label}
\usage{
get_piece(
  labels,
  piece = "all",
  notation = RCLabels::bracket_notation,
  prepositions = RCLabels::prepositions
)
}
\arguments{
\item{labels}{The row and column labels from which prepositional phrases are to be extracted.}

\item{notation}{The notation object that describes the labels.
Default is \code{RCLabels::bracket_notation}.}

\item{prepositions}{A vector of strings to be treated as prepositions.
Note that a space is appended to each word internally,
so, e.g., "to" becomes "to ".
Default is \code{RCLabels::prepositions}.}

\item{which}{The name of the item to return.}
}
\value{
A \code{piece} of \code{labels}.
}
\description{
This is a wrapper function for \code{keep_pref_suff()}, \code{get_nouns()}, and
\code{get_objects()}.
It returns a \code{piece} of a row or column label.
}
\details{
\code{piece} is typically one of
\itemize{
\item "all" (which returns \code{labels} directly),
\item "pref" (for the prefixes),
\item "suff" (for the suffixes),
\item "noun" (returns the noun) or
\item "<<a preposition>>" which will return the object of <<a preposition>>
in \code{labels}.
}

\code{piece} must be a character vector of length 1.

If a \code{piece} is missing in a label, "" (empty string) is returned.
}
