% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_pps}
\alias{get_pps}
\title{Extract prepositional phrases from labels}
\usage{
get_pps(
  labels,
  notation = RCLabels::bracket_notation,
  prepositions = RCLabels::prepositions
)
}
\arguments{
\item{labels}{The row and column labels from which prepositional phrases are to be extracted.}

\item{notation}{The notation object that describes the labels.
Default is \code{RCLabels::bracket_notation}.}
}
\value{
A list of lists of prepositional phrases.
The length of the list is the same as the length of \code{labels}.
}
\description{
A suffix can consist of several prepositional phrases.
This function splits the prepositional phrases apart,
returning a list of those phrases.
Each prepositional phrase consists of a proposition
(see \code{RCLabels::prepositions}) and its object.
}
\details{
Some labels will have no suffix and no prepositions.
When there are no prepositions, \code{NA} is returned.
}
\examples{
get_pps(c("a [of b in c]", "d [of e into f]"), bracket_notation)
}
