% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_pps_old}
\alias{get_pps_old}
\title{Extract prepositional phrases from row and column labels}
\usage{
get_pps_old(
  labels,
  notation = RCLabels::bracket_notation,
  prepositions = RCLabels::prepositions
)
}
\arguments{
\item{labels}{The row and column labels from which prepositional phrases are to be extracted.}

\item{notation}{The notation object that describes the labels.
Default is \code{RCLabels::bracket_notation}.}

\item{prepositions}{A vector of strings to be treated as prepositions.
Note that a space is appended to each word internally,
so, e.g., "to" becomes "to ".
Default is \code{RCLabels::prepositions}.}
}
\value{
A list of lists of prepositional phrases.
The structure of the returned list is the same as the length of \code{labels}
with an additional level added
in which each item is one prepositional phrase.
}
\description{
A suffix can consist of several prepositional phrases.
This function returns the prepositional phrases as a single string.
Each prepositional phrase consists of a proposition
(see \code{RCLabels::prepositions}) and its object.
}
\details{
Some labels will have no suffix and no prepositions.
When there are no prepositions, \code{NA} is returned.
}
\examples{
get_pps(c("a [of b in c]", "d [of e into f]"), bracket_notation)
}
