% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{in_notation}
\alias{in_notation}
\title{In notation}
\format{
A vector of notational symbols that provides to ("a [in b]") notation.
}
\usage{
in_notation
}
\description{
A description of in notation.
}
\examples{
in_notation
}
\keyword{datasets}
