% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{infer_notation_for_one_label}
\alias{infer_notation_for_one_label}
\title{Infer the notation from one row or column label}
\usage{
infer_notation_for_one_label(
  x,
  notations = RCLabels::notations_list,
  allow_multiple = FALSE,
  retain_names = FALSE,
  choose_most_specific = FALSE
)
}
\arguments{
\item{x}{A single row or column label.}

\item{notations}{The notations from which matches can be discerned.
Default is \code{RCLabels::notations_list}.}

\item{allow_multiple}{A boolean that tells whether multiple notation matches
are allowed.
If \code{FALSE} (the default), multiple matches give an error.}

\item{retain_names}{A boolean that tells whether to retain names on the
outgoing matches.
Default is \code{FALSE}.
If \code{TRUE}, the return value is a named list.
If only one of \code{notations} is returned,
names are never supplied.}

\item{choose_most_specific}{A boolean that indicates if the most-specific notation
will be returned when more than one of \code{notations} matches \code{x}.
Most specific is defined as the notation with the most characters.
If two or more notations with the same number of characters match,
the first matching notation in \code{notations} is returned.
Default is \code{FALSE}.
When \code{choose_most_specific = TRUE},
the values of \code{allow_multiple} and \code{retain_names} no longer matter.
At most one of the \code{notations} will be returned without names.}
}
\value{
A single matching notation object (if \code{allow_multiple = FALSE}, the default)
or possibly multiple matching notation objects (if \code{allow_multiple = TRUE}).
If no \code{notations} match \code{x}, \code{NULL}.
}
\description{
This is a non-public helper function for vectorized \code{infer_notation()}.
}
