% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{match_notation}
\alias{match_notation}
\title{Tells which notations match a row or column label}
\usage{
match_notation(x, notations = RCLabels::notations_list, allow_multiple = FALSE)
}
\arguments{
\item{x}{A row or column label (or vector of labels).}

\item{notations}{The notations from which matches can be found.
Default is \code{RCLabels::notations_list}.}

\item{allow_multiple}{A boolean that tells whether multiple matches are allowed.
If \code{FALSE} (the default), multiple matches give an error.}
}
\value{
A single notation object (if \code{x} is a single row or column label)
or a vector of notation objects (if \code{x} is a vector).
If no \code{notations} match \code{x}, \code{NULL} is returned.
}
\description{
It is convenient to know which notation is applicable to a row or column label.
This function tells which \code{notations} are appropriate for \code{x}.
}
\details{
By default (\code{allow_multiple = FALSE}),
a single notation object is returned if only one notation in \code{notations}
is appropriate for \code{x}.
Multiple matches can be returned if \code{allow_multiple} is set \code{TRUE}.

\code{x} can be a vector, in which case the output is also a vector of notations.
}
