% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_pattern}
\alias{match_pattern}
\title{Tell whether row or column labels match a regular expression}
\usage{
match_pattern(
  labels,
  regex_pattern,
  pieces = "all",
  prepositions = RCLabels::prepositions,
  notation = RCLabels::bracket_notation,
  ...
)
}
\arguments{
\item{labels}{The row and column labels in which pieces will be modified.}

\item{regex_pattern}{The regular expression pattern to determine matches.
Consider using \code{Hmisc::escapeRegex()} to escape \code{regex_pattern}
before calling this function.}

\item{notation}{The notation used in \code{labels}.
Default is \code{RCLabels::bracket_notation}.}

\item{...}{Other arguments passed to \code{grepl()}, such as \code{ignore.case}, \code{perl}, \code{fixed},
or \code{useBytes}.
See examples.}

\item{piece}{The piece (or pieces) of row or column labels to be checked for matches.
See details.}
}
\value{
A logical vector of same length as \code{labels},
where \code{TRUE} indicates a match was found and \code{FALSE} indicates otherwise.
}
\description{
This function tells whether row or column labels
match a regular expression.
Internally, \code{grepl()} decides whether a match occurs.
}
\details{
By default (\verb{pieces = "all}), complete labels (as strings) are checked for matches.
If \code{pieces == "pref"} or \code{pieces == "suff"},
only the prefix or the suffix is checked for matches.
Alternatively, \code{pieces = "noun"} or \verb{pieces = <<preposition>>} indicate
that only specific pieces of labels are to be checked for matches.
}
