% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{of_notation}
\alias{of_notation}
\title{Of notation}
\format{
A vector of notational symbols that provides of ("a [of b]") notation.
}
\usage{
of_notation
}
\description{
A description of of notation.
}
\examples{
of_notation
}
\keyword{datasets}
