% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paren_notation}
\alias{paren_notation}
\title{Parenthetical notation}
\format{
A vector of notational symbols that provides a parenthetical ("a (b)") notation.
}
\usage{
paren_notation
}
\description{
A description of parenthetical notation.
}
\examples{
paren_notation
}
\keyword{datasets}
