% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prepositions}
\alias{prepositions}
\title{Prepositions}
\format{
A vector of prepositions used in row and column labels.
}
\usage{
prepositions
}
\description{
This constant is deprecated.
Please use \code{prepositiions_list} instead.
}
\keyword{datasets}
