% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prepositions_list}
\alias{prepositions_list}
\title{Prepositions}
\format{
A vector of prepositions used in row and column labels.
}
\usage{
prepositions_list
}
\description{
Prepositions used in row and column labels.
}
\examples{
prepositions_list
}
\keyword{datasets}
