% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{recombine_labels}
\alias{recombine_labels}
\title{Recombine row and column labels}
\usage{
recombine_labels(ls, notation = RCLabels::bracket_notation)
}
\arguments{
\item{ls}{A vector of split row or column labels.}

\item{notation}{The notation object that describes the labels.
Default is \code{RCLabels::bracket_notation}.}
}
\value{
Recombined row and column labels.
}
\description{
This function recobines (unsplits) row or column labels that have
been separated by \code{split_labels()}.
}
\examples{
labs <- c("a [of b in c]", "d [from Coal mines in USA]")
split <- split_labels(labs)
split
recombine_labels(split)
}
