% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_nouns}
\alias{set_nouns}
\title{Set nouns in labels}
\usage{
set_nouns(labels, new_nouns, notation = RCLabels::bracket_notation)
}
\arguments{
\item{labels}{The labels in which the nouns will be set.}

\item{new_nouns}{The new nouns to be set in \code{labels}.
Must be same length as \code{labels}.}

\item{notation}{The notation used in \code{labels}.
Default is \code{RCLabels::bracket_notation}.}
}
\value{
A character vector of same length as labels
with \code{new_nouns}.
}
\description{
This function sets the nouns of row and column labels.
The length of \code{new_nouns} must be the same as the length of \code{labels}.
}
\examples{
labels <- c("a [of b in c]", "d [of e in USA]")
set_nouns(labels, c("a_plus", "g"))
}
