% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{split_labels}
\alias{split_labels}
\title{Split row and column labels into nouns and prepositional phrases}
\usage{
split_labels(
  labels,
  notation = RCLabels::bracket_notation,
  prepositions = RCLabels::prepositions
)
}
\arguments{
\item{labels}{The row and column labels from which prepositional phrases are to be extracted.}

\item{notation}{The notation object that describes the labels.
Default is \code{RCLabels::bracket_notation}.}

\item{prepositions}{A vector of strings to be treated as prepositions.
Note that a space is appended to each word internally,
so, e.g., "to" becomes "to ".
Default is \code{RCLabels::prepositions}.}
}
\value{
A list of lists with items named \code{noun} and \code{pp}.
}
\description{
Split row and column labels into nouns and prepositional phrases
}
\examples{
split_labels(c("a [of b in c]", "d [of e into f]"),
             notation = bracket_notation)
}
