% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{strip_label_part}
\alias{strip_label_part}
\title{A convenience function to help splitting prefixes and suffixes}
\usage{
strip_label_part(x, notation, part, pattern_pref = "", pattern_suff = "")
}
\arguments{
\item{x}{The label(s) to be split.}

\item{notation}{The notations to be used for each \code{x}.}

\item{part}{The part of the label to work on, such as "pref_start", "pref_end", "suff_start", or "suff_end".}

\item{pattern_pref}{The prefix to a regex pattern to be used in \code{gsub()}.}

\item{pattern_suff}{The suffix to a regex pattern to be used in \code{gsub()}.}
}
\value{
A label shorn of the part to be stripped.
}
\description{
This function should only ever see a single label (\code{x})
and a single \code{notation}.
}
\details{
If \code{notation} is \code{NULL}, \code{x} is returned, unmodified.
}
