% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{to_notation}
\alias{to_notation}
\title{To notation}
\format{
A vector of notational symbols that provides to ("a [to b]") notation.
}
\usage{
to_notation
}
\description{
A description of to notation.
}
\examples{
from_notation
}
\keyword{datasets}
