% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_pps}
\alias{get_pps}
\title{Extract prepositional phrases of row and column labels}
\usage{
get_pps(
  labels,
  notation = RCLabels::notations_list,
  prepositions = RCLabels::prepositions_list,
  choose_most_specific = TRUE
)
}
\arguments{
\item{labels}{A list or vector of labels from which prepositional phrases are to be extracted.}

\item{notation}{The notation type to be used when extracting prepositional phrases.
Default is \code{RCLabels::bracket_notation}.}

\item{prepositions}{A list of prepositions, used to detect prepositional phrases.
Default is \code{RCLabels::bracken_notation}.}

\item{choose_most_specific}{A boolean that tells whether to choose the most specific
\code{notation} if more than 1 match is found.}
}
\value{
All prepositional phrases in a suffix.
}
\description{
This function extracts prepositional phrases from suffixes of row and column labels
of the form "a \link{preposition b}", where "preposition b" is the prepositional phrase.
}
\examples{
get_pps(c("a [in b]", "c [of d]"))
get_pps(c("a [of b in c]", "d [-> e of f]"))
}
