% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notations_list}
\alias{notations_list}
\title{Notations list}
\format{
A list of bundled notations.
}
\usage{
notations_list
}
\description{
A list of all bundled notations.
This list is organized by least specific to most specific,
thereby enabling some unique behaviors in \code{infer_notation()}.
See the examples for \code{infer_notation()}.
}
\examples{
notations_list
}
\keyword{datasets}
